♻️ refactord: 간단한 패키지 이동 및 피드백 구현
