+++ src/main/java/com/sparta/sensorypeople/domain/user/dto/RefreshTokenRequestDto.java
@@ -1,5 +1,6 @@
 package com.sparta.sensorypeople.domain.user.dto;
 
+import jakarta.validation.constraints.NotBlank;
 import lombok.Getter;
 
 /**
@@ -12,5 +13,6 @@ public class RefreshTokenRequestDto {
     /**
      * 리프레시 토큰
      */
+    @NotBlank(message = "리프레시 토큰을 입력해주세요")
     private String refreshToken;
 }
