+++ src/main/java/com/sparta/sensorypeople/security/service/UserDetailsServiceImpl.java
@@ -1,4 +1,4 @@
-package com.sparta.sensorypeople.domain.user.service;
+package com.sparta.sensorypeople.security.service;
 
 import com.sparta.sensorypeople.common.exception.CustomException;
 import com.sparta.sensorypeople.common.exception.ErrorCode;
